pyproject.toml
[project]
name = "deckbot"
version = "0.1.0"
description = "MYSTRAN deck-archiving Discord bot"
requires-python = ">=3.12"
dependencies = [
  "discord.py>=2.3",
  "sqlalchemy[asyncio]>=2.0",
  "aiosqlite>=0.19",
  "alembic>=1.13",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "zstandard>=0.22",
]

[project.scripts]
deckbot = "deckbot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["deckbot"]

[tool.ruff]
target-version = "py314"
indent-width = 2
line-length = 79

[tool.ruff.lint.flake8-quotes]
avoid-escape = false

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
quote-style = "double"

[tool.ruff.lint]
extend-select = ["I"]
extend-fixable = ["I"]
